[PDS-74] refactor: resource 폴더명 수정
